+++ src/path/filepath/example_unix_test.go
@@ -8,7 +8,6 @@ package filepath_test
 
 import (
 	"fmt"
-	"os"
 	"path/filepath"
 )
 
@@ -79,26 +78,4 @@ func ExampleJoin() {
 	// a/b/c
 	// a/b/c
 	// a/b/c
-}
-
-func ExampleWalk() {
-	dir := "dir/to/walk"
-	subDirToSkip := "skip" // dir/to/walk/skip
-
-	err := filepath.Walk(dir, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			fmt.Printf("prevent panic by handling failure accessing a path %q: %v\n", path, err)
-			return err
-		}
-		if info.IsDir() && info.Name() == subDirToSkip {
-			fmt.Printf("skipping a dir without errors: %+v \n", info.Name())
-			return filepath.SkipDir
-		}
-		fmt.Printf("visited file: %q\n", path)
-		return nil
-	})
-
-	if err != nil {
-		fmt.Printf("error walking the path %q: %v\n", dir, err)
-	}
 }
